+++ app/services/bot_context/handle_service.rb
@@ -6,11 +6,14 @@ module BotContext
       telegram_api: 'api.telegram.client',
       handle_command: 'services.bot_context.handle_command',
       represent_command: 'services.bot_context.represent_command',
+      represent_raw_command: 'services.bot_context.represent_raw_command',
       handle_webhook: 'services.webhooks_context.telegram.handle_message_webhook'
     ]
 
     SERVICE_COMMANDS = %w[/start /contacts /unsubscribe /subscribe /help /commands].freeze
     SERVICE_SOURCES = %i[telegram_bot].freeze
+    TELEGRAM_SOURCES = %i[telegram_bot telegram_group_bot].freeze
+    RESPONSE_SOURCES = %i[web raw].freeze
 
     # rubocop: disable Metrics/AbcSize
     def call(source:, message:, data: {})
@@ -54,18 +57,47 @@ module BotContext
     # rubocop: enable Style/RedundantRegexpArgument
 
     def response(source, result, data={})
-      return send_result_message(data[:raw_message], result) if source != :web && source != :raw
-
-      { result: result[:result], errors: result[:errors], errors_list: result[:errors] }
+      send_result_message_in_response(data[:raw_message], result) if source.in?(TELEGRAM_SOURCES)
+      send_messages_to_channels(result, data) if source == :raw && data[:character]
+      { result: result[:result], errors: result[:errors], errors_list: result[:errors] } if source.in?(RESPONSE_SOURCES)
     end
 
-    def send_result_message(raw_message, command_formatted_result)
+    def send_messages_to_channels(result, data)
+      command_result = result.merge(character: data[:character])
+      data[:character].channels.find_each do |channel|
+        command_formatted_result =
+          represent_raw_command.call(
+            source: channel.provider.to_sym,
+            command: '/check',
+            provider: character_provider(data[:character].class.name),
+            command_result: command_result
+          )
+        send_result_message(channel.external_id, command_formatted_result)
+      end
+    end
+
+    def send_result_message_in_response(raw_message, command_formatted_result)
       telegram_api.send_message(
         bot_secret: bot_secret,
         chat_id: raw_message.dig(:chat, :id),
         reply_to_message_id: raw_message[:message_id],
         text: command_formatted_result[:errors] ? command_formatted_result.dig(:errors, 0) : command_formatted_result[:result]
       )
+    end
+
+    def send_result_message(external_id, command_formatted_result)
+      telegram_api.send_message(
+        bot_secret: bot_secret,
+        chat_id: external_id,
+        text: command_formatted_result[:errors] ? command_formatted_result.dig(:errors, 0) : command_formatted_result[:result]
+      )
+    end
+
+    def character_provider(name)
+      case name
+      when 'Dnd5::Character', 'Dnd2024::Character' then 'dnd'
+      when 'Daggerheart::Character' then 'daggerheart'
+      end
     end
 
     def bot_secret